+++ tests/unit/metadata/test_multi_table.py
@@ -555,6 +555,24 @@ class TestMultiTableMetadata:
         # Assert
         assert set(result) == {'user_id', 'session_id'}
 
+    def test__get_all_keys(self):
+        """Test that this method returns all key columns for a table."""
+        # Setup
+        instance = self.get_metadata()
+        instance.add_column('transactions', 'user_id', sdtype='id')
+        instance.add_relationship(
+            parent_table_name='users',
+            parent_primary_key='id',
+            child_table_name='transactions',
+            child_foreign_key='user_id',
+        )
+
+        # Run
+        result = instance._get_all_keys('transactions')
+
+        # Assert
+        assert set(result) == {'user_id', 'session_id', 'transaction_id'}
+
     def test_add_relationship(self):
         """Test the ``add_relationship`` method of ``MultiTableMetadata``.
 
@@ -3337,3 +3355,35 @@ class TestMultiTableMetadata:
         # Run and Assert
         with pytest.raises(InvalidMetadataError, match=error_msg):
             metadata.validate_data(data)
+
+    @patch('sdv.metadata.multi_table.create_summarized_columns_node')
+    @patch('sdv.metadata.multi_table.create_columns_node')
+    def test__get_table_info(self, mock_columns_node, mock_summarized_columns_node):
+        """Test that the `_get_table_info` method."""
+        # Setup
+        mock_columns_node.return_value = 'column'
+        mock_summarized_columns_node.return_value = 'column'
+        metadata = MultiTableMetadata()
+        table = Mock()
+        table.primary_key = 'primary_key'
+        table.sequence_key = None
+        table.sequence_index = None
+        table_all = Mock()
+        table_all.primary_key = 'primary_key'
+        table_all.sequence_index = 'sequence_index'
+        table_all.sequence_key = 'sequence_key'
+        metadata.tables = {'table': table, 'table_all': table_all}
+
+        # Run
+        result = metadata._get_table_info('table', show_table_details='full')
+        result_all = metadata._get_table_info('table_all', show_table_details='summarized')
+        result_no_info = metadata._get_table_info('table_all', show_table_details=None)
+
+        # Assert
+        assert result['primary_key'] == 'Primary key: primary_key'
+        assert 'sequence_index' not in result
+        assert 'sequence_key' not in result
+        assert result_all['primary_key'] == 'Primary key: primary_key'
+        assert result_all['sequence_key'] == 'Sequence key: sequence_key'
+        assert result_all['sequence_index'] == 'Sequence index: sequence_index'
+        assert result_no_info is None